+++ examples/transformer_app.py
@@ -49,5 +49,5 @@ optimizer = fleet.distributed_optimizer(optimizer)
 optimizer.minimize(model.loss)
 
 place = fluid.CUDAPlace(int(os.environ.get('FLAGS_selected_gpus', 0)))
-trainer = X.Trainer(place)
+trainer = X.MultiGPUTrainer(place)
 trainer.fit(model, data_loader, epoch=10)
