+++ python/fleetx/__init__.py
@@ -17,4 +17,6 @@ from . import applications
 from .applications import *
 from . import dataset
 from .dataset import *
-from .applications.trainer import Trainer
+from .applications.trainer import MultiGPUTrainer, CPUTrainer
+from . import utils
+from .utils import *
